First upload
Starting work on this, some main bits are being added but not being
checked just yet
* path finding has been added, to find lengths to the first materials
sections
* neutron path is in 3D (r,z,n) coordinates and related to blankets in
2D
* I need to change the method this uses, as it currently can only give
length, but can't tell what material is there
This is specifically the pathintersect function. It should be changed to
return a length and material list, which means that the intersections
need to be done for each volume present - if there is an intersection
there, then the points of intersection must found
Notes:
There must be an even number of points of intersection with each blanket
module that is not the starting module, as the line extends to infinity
from the origin.
Length must be calculated from the 3D line,